+++ docs/changelog.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
+***Fixed:***
+
+- Properly handle failed downloads
+
 ## 0.20.0 - 2024-05-13
 
 ***Added:***